// frontend/src/api/gen/models/noteVisibility.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Boardit API
 * API Contract – v1
 *
 * Status: Stable (v1)
 * Scope: Auth & User (legacy paths) + Notes/Folder Tree/Autosave/XSS
 *
 * OpenAPI spec version: 1.0.0
 */

export type NoteVisibility = typeof NoteVisibility[keyof typeof NoteVisibility];


export const NoteVisibility = {
  private: 'private',
  unlisted: 'unlisted',
  public: 'public',
} as const;